+++ pkg/step/step.go
@@ -25,7 +25,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/joho/godotenv"
 	"github.com/mattn/go-sixel"
 	pb "sqlflow.org/sqlflow/pkg/proto"
 	"sqlflow.org/sqlflow/pkg/sql"
@@ -155,12 +154,4 @@ func GetStdout(f func() error) (out string, e error) {
 	os.Stdout, os.Stderr = oldStdout, oldStderr // restoring the real stdout and stderr
 	out = <-outC
 	return
-}
-
-// DotEnvFilename is the filename of the .env file
-const DotEnvFilename string = ".sqlflow_env"
-
-// InitEnvFromFile initializes environment variables from the .env file
-func InitEnvFromFile(f string) {
-	_ = godotenv.Load(f)
 }
